Fix numeric generators, add default value tests

+++ Faker.Core/Generators/ByteGenerator.cs
@@ -7,7 +7,7 @@ namespace Faker.Core.Generators
     {
         public object Generate(Type typeToGenerate, GeneratorContext context)
         {
-            return context.Random.Next(1, byte.MaxValue);
+            return (byte)context.Random.Next(1, byte.MaxValue);
         }
 
         public bool CanGenerate(Type type)
