+++ Faker.Core/Generators/ShortGenerator.cs
@@ -7,7 +7,7 @@ namespace Faker.Core.Generators
     {
         public object Generate(Type typeToGenerate, GeneratorContext context)
         {
-            return context.Random.Next(1, short.MaxValue);
+            return (short)context.Random.Next(1, short.MaxValue);
         }
 
         public bool CanGenerate(Type type)
